flaskr/requirements.txt
alembic==1.13.3
aniso8601==9.0.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
decorator==4.4.2
Faker==30.6.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google_cloud_storage==2.18.2
greenlet==3.1.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.23.0
marshmallow-sqlalchemy==1.1.0
moviepy==1.0.3
numpy==1.26.4
packaging==24.1
Pillow==9.5.0
proglog==0.1.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pytz==2024.2
redis==5.2.0
requests==2.32.3
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4